+++ example/menu_provider_example/lib/menu_provider.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class MenuProvider extends ChangeNotifier {
+  List<PlatformMenu> _currentMenus = [];
+
+  List<PlatformMenu> get currentMenus => _currentMenus;
+
+  void setMenus(List<PlatformMenu> menus) {
+    _currentMenus = menus;
+    notifyListeners();
+  }
+}